UPSTREAM: mmc: core: Enable / disable re-tuning
Enable re-tuning when tuning is executed and
disable re-tuning when card is no longer initialized.

In the case of SDIO suspend, the card can keep power.
In that case, re-tuning need not be disabled, but, if
a re-tuning timer is being used, ensure it is disabled
and assume that re-tuning will be needed upon resume
since it is not known how long the suspend will last.

Signed-off-by: Adrian Hunter <adrian.hunter@intel.com>
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
(cherry picked from commit 79d5a65aeea43920bf3ff60791f317570dd6f54f)
Signed-off-by: Marc Herbert <marc.herbert@intel.com>

BUG=chrome-os-partner:40410
TEST=With this whole HS400 series applied:
TEST=on a system with HS400 controller(s), it/they initialize with the
  following log message:
  INFO kernel: [  103.601747] mmc0: new HS400 MMC card at address 0001
TEST=run SD/eMMC test plan: higher eMMC throughput after change, and
  no other change.
TEST=same on another system with no HS400: no observable change at all.

Change-Id: I8e95c2c8604a7002fa79164a5160c7cbd6fbc82d
Reviewed-on: https://chromium-review.googlesource.com/276400
Commit-Ready: Marc Herbert <marc.herbert@intel.com>
Tested-by: Philip Hanson <philip.hanson@intel.com>
Reviewed-by: Derek Basehore <dbasehore@chromium.org>