BACKPORT: tytso: ext4 crypto: optimize filename encryption
Encrypt the filename as soon it is passed in by the user.  This avoids
our needing to encrypt the filename 2 or 3 times while in the process
of creating a filename.

Similarly, when looking up a directory entry, encrypt the filename
early, or if the encryption key is not available, base-64 decode the
file syystem so that the hash value and the last 16 bytes of the
encrypted filename is available in the new struct ext4_filename data
structure.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>
(cherry picked from commit 5b643f9ce34df945e58c7176275d406aa0db704f)

BUG=chromium:347322
TEST=Compiled, Installed on samus
TEST=Test with e2fsprogs 1.43-WIP

Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Change-Id: I627fffef25c13ccf21cc5c0d2115b88d54cdd70b
Previous-Reviewed-on: https://chromium-review.googlesource.com/321687
(cherry picked from commit 35deca5305966435c75d54e7cac0ec1b964320e8)
Reviewed-on: https://chromium-review.googlesource.com/334196