CHROMIUM: extcon: Introduce cros_ec extcon driver
This is the driver for the USB Type C cable detection mechanism
built into the ChromeOS Embedded Controller on systems that
have USB Type-C ports.

At present, this allows for the detection of host and device
type cables, but in the future, it may also be used to notify
the presence of power and display out as well.

Signed-off-by: Benson Leung <bleung@chromium.org>

BUG=chromium:488161,chrome-os-partner:38897
TEST=Build and test on smaug.
Plug in a USB-A dongle and a usb device such as a ethernet adapter.
cat /sys/class/extcon/extcon-cros-ec.1/state
Check that the following appears:
USB=0
USB-Host=1

Unplug the cable. Cat state again :
USB=0
USB-Host=0

This time, plug in a A-C cable from a host PC (act as gadget).
USB=1
USB-Host=0

Change-Id: I81fc41bb00cfbe55bdb39b01521f1ccd2d9f62f1
Reviewed-on: https://chromium-review.googlesource.com/272471
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Commit-Queue: Benson Leung <bleung@chromium.org>
Trybot-Ready: Benson Leung <bleung@chromium.org>
Tested-by: Benson Leung <bleung@chromium.org>