Revert "CHROMIUM: arm64: add option to build Image.fit fit image"
This reverts commit 1ce69abbf412da871c6aa461a34e71c51e471db6.

We only produce Image as part of the native kernel environment, and wrap the
kernel in an appropriate format as part of the build system instead.

Upstream has taken a hard stance on not allowing custom image production
as part of the kernel build, so keeping the same policy in our downstream
project helps reduce confusion and makes our build environment agnostic
of where the kernel tree is coming from (and whether it has our local
changes to directly produce the Image.fit).

We are for now keeping the generate-its-script.sh shell script around in
our tree, but that should at some point move out into the surrounding
build environment as well.

Change-Id: Id3c14ff263ed2a1d8bd44de47d4049ee6d2338f8
Signed-off-by: Olof Johansson <olofj@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/272909
Reviewed-by: David Riley <davidriley@google.com>